Removing Raw Input mode for now, sticking the basics until I understand it more
