+++ src/level1/level1.rego
@@ -3,6 +3,6 @@ package barista
 
 # Create the rule below
 deny[message] {
-    input.Order.Drink == "Cola"
+    input.Order.Drink == "Beer"
     message := sprintf("Unfortunately, we do not serve %s", [input.Order.Drink])
 }
