+++ src/level1/mod.rs
@@ -4,33 +4,17 @@ use std::fs;
 use std::io;
 use std::process::Command;
 
-use self::termion::event::Key;
-use self::termion::input::TermRead;
-use self::termion::raw::IntoRawMode;
-
 pub fn run() -> Result<bool, String> {
-    print!("{}{}", termion::cursor::Goto(1, 1), termion::clear::All);
-    println!(
-        "It's the year 3000, you are the latest and greatest robot barista. In the year 3000 all orders are fed to you through a JSON file, its then up to you if the order is valid or not. Luckily you've been modded to be able to use OPA (Open Policy Agent) which will help you set some rules on the fly (nice!).\n"
-    );
+    println!("It's the year 3000, you are the latest and greatest robot barista. In the year 3000 all orders are fed to you through a JSON file, its then up to you if the order is valid or not. Luckily you've been modded to be able to use OPA (Open Policy Agent) which will help you set some rules.");
     println!("Press Enter if you are ready!");
-
     let mut input = String::new();
     io::stdin().read_line(&mut input).expect("Failed to read line");
 
-    let mut success = false;
-    
-
     loop {
-        print!("{}{}", termion::cursor::Goto(1, 1), termion::clear::All);
-        println!(
-            "As the proud employee of a specialty coffee shop, you can't believe your robot ears when a customer walks in and asks for a cola, completely disregarding the carefully crafted menu of premium coffee offerings."
-        );
-        println!("\nLevel 1 Order (level1.json):\n");
+        println!("As the proud employee of a specialty coffee shop, you can't believe your robot ears when a customer walks in and asks for a cola, completely disregarding the carefully crafted menu of premium coffee offerings.");
+        println!("Level 1 Order (level1.json):");
 
-        let file_contents = fs
-            ::read_to_string("src/level1/level1.json")
-            .expect("Failed to read level1.json");
+        let file_contents = fs::read_to_string("src/level1/level1.json").expect("Failed to read level1.json");
 
         let output = Command::new("opa")
             .arg("eval")
@@ -67,32 +51,20 @@ pub fn run() -> Result<bool, String> {
 
         if stdout.contains("Unfortunately, we do not serve Cola") {
             println!("Well done! You did it!");
-            success = true;
-            return Ok(success);
+            return Ok(true);
         } else {
-            println!("\nWhat would you like to do?");
+            println!("What would you like to do?");
             println!("1. Retry");
             println!("2. Exit");
 
-            // Enter raw mode so that we can read single keystrokes
-            let stdout = io::stdout().into_raw_mode().unwrap();
-            let stdin = io::stdin();
-            let reader = stdin.keys();
+            let mut input = String::new();
+            io::stdin().read_line(&mut input).expect("Failed to read line");
 
-            for key in reader {
-                match key.unwrap() {
-                    Key::Char('1') => {
-                        break;
-                    }
-                    Key::Char('2') => {
-                        return Ok(false);
-                    }
-                    _ => {}
-                }
+            match input.trim() {
+                "1" => {},
+                "2" => return Ok(false),
+                _ => {}
             }
-
-            // Leave raw mode before exiting
-            drop(stdout);
         }
     }
 }
